Fix up scrolling label - ensure return to left margin From Martin Williams <martinr.williams@gmail.com>
On a Squeezebox Controller, some scrolling labels suffer from a
"jittering" issue, when they will bounce a small amount back and forth.

The underlying issue is also present on a Squeezebox Radio, and any
version of Squeezeplay, although I have not seen the "jittering" that a
Controller suffers. Nevertheless the effect can also be seen, in that a
label may end up slightly offset from the left margin after scrolling is
stopped.

The cause is a defect in the label animation rendering code, whose
effect is that when a label reaches the end of its scrolling period, it
is not redrawn at its "home" position, and may be left at an offset of
a few pixels from the left margin.

The problem was introduced within this commit:
  return to JIVE_FRAME_RATE / 2 frame rate
  https://github.com/Logitech/squeezeplay/commit/0c09be102c0e2971fb2b79a68bb1df6d9270c12b

This optimization is the cause:
        if (peer->scroll_offset < 0) {
                return 0;
        }

Prior to that change, the label would have been redrawn in its correct
position on every animation update. Perhaps rather too often.

But, now, the label is not redrawn if the scrolling period is stopped.
Fine, but there is no guarantee that the label will have been last drawn
at the left margin. And it often isn't, hence the problem.