Bugfix: Incomplete extraction of icy metadata causes Squeezeplay to quit From Martin Williams <martinr.williams@gmail.com>
Extracting icy metadata from an incoming stream requires two data reads
from the fifo circular stream buffer if the icy metadata should roll
around the end of the buffer.

At present only one read is attempted, and the code assumes that a short
read implies a fundamental failure in the integrity of the stream
buffer. [assert(r == icy_len)].

The consequence of the assertion failure is that Squeezeplay will simply
quit.

This change corrects matters by undertaking two reads if the initial
read is 'short'. This behaviour is mandated by the implementation of
'streambuf_fast_read' and the fifo circular stream buffer.

We know that there are sufficient bytes available to satisfy the full
read. So the only valid state after the first 'short' read is that the
fifo read buffer pointer is set at the start of the buffer. We implement
an assertion check on that condition. It assures integrity of the stream
buffer, and is in the spirit of the existing code.