Wi-fi set up: Suppress hidden SSIDs under wpa_supplicant >= v2.0
Modern versions of wpa_supplicant have changed the way in which SSIDs
are displayed. The change allows all bytes in a SSID to be reported
accurately, by making use of a hex-escape encoding scheme.

A side effect of this is that "hidden" SSIDs may now become visible. For
example: "\x00\x00\x00\x00\x00\x00\x00\x00\x00" is how a SSID composed
of nine null bytes is reported in the new scheme.

This doesn't look good, and will be confusing to the user. SqueezePlay
can make no use of such SSIDs to make a connection.

This change addresses the issue by adapting '_scanResults' in
'SetupNetworkingApplet.lua' to discard any SSIDs that start with a null
byte.

Under earlier versions of wpa_supplicant, these SSIDs would have been
returned as null strings. '_wirelessScanTask' in 'Networking.lua' skips
over null strings, and they would not show up.

This change makes no attempt to determine which version of
wpa_supplicant is running. It is possible that a user might have
actually set a SSID which starts with the character sequence "\x00". In
that case, the Touch and Controller, which run an older wpa_supplicant,
will no longer include the SSID in their listings. This is unlikely to
occur in practice. Manual entry of the SSID is still possible.

We should expect to see more use of this feature in future. For example,
consumer access points may well offer the ability to use utf-8 encoding
for a SSID. At present, SqueezePlay would not be able to make use of
such a SSID. The point may need to be addressed in the future.

Further remarks:

On or around 7 August 2012 wpa_supplicant was modified to report SSIDs
in a printf-escaped format.

Refer, for example, commit 6bc1f95 in git://w1.fi/srv/git/hostap.git:
Use printf escaping in SSID-to-printable-string conversion

Now, rather than simply replacing non-ASCII characters with underscores,
hex escaped strings are used. Furthermore, a SSID containing null bytes
will no longer be simply truncated at the first null byte.