Bugfix: Manually repairing network may lock up UI on failure
This change hardens up the "Repair network" option in the SqueezeOS
diagnostics menu, under "Network Health".

Taking the "Repair network" option will lock up the UI if, for any
reason, 'Networking:repairNetwork' does not properly terminate. This may
occur if it is called without a valid network interface, or if, for any
reason, the Wi-Fi system fails to respond as expected.

The user will be left with a "spinny" style connection progress screen,
without any means of backing out.

The change implements the following:

+ "Repair network" will now display a completion status message in the
UI, matching the behaviour of the "Check Network" option.

+ 'Networking:repairNetwork' now checks that the active network
interface is defined, and returns an appropriate error if not.

+ A backstop timer of 20 seconds is implemented that, if necessary, will
close the connection progress spinny and show a failed completion
status. I believe that 20 seconds should be more than sufficient time
for a repair to succeed, but experience may dictate otherwise. Note that
this does not terminate 'Networking:repairNetwork', so the repair may
yet succeed. (And will update the completion status accordingly).

Remarks:

'Networking:repairNetwork' itself launches a 'repairnetwork' task. The
operations carried out in that task, 'Networking:_ifDown' and
'Networking:_ifUp', both contain a number of 'assert' statements that,
if they fail, will cause improper termination. Ideally, what is wanted
is that failures be captured robustly at this level. It may be possible
to achieve this by revisiting and hardening up some of the underlying
processes, but that will require further study and testing. The primary
purpose of this change is to prevent UI lock up.