Timezone changes - GMT-11:00 - Midway Island & Samoa
This zone was represented by "Pacific/Apia", the capital of Western
Samoa. But Western Samoa "crossed the date line" in 2011, so it is no
longer applicable to American Samoa and Midway.

The zone is now represented by "Pacific/Pago_Pago", the capital of
American Samoa. DST is not observed.

Apia (Western Samoa) has been added to the pre-existing GMT+13:00 zone
for Nuku'alofa (Tonga). DST is not observed in either location.