LibSDL/Squeezeplay - Fix up build failure cross compiling SDL client programs
This change takes up a fix to SDL's Automake macro to ensure that, when
cross compiling a client program, the correct version of 'sdl-config' is
used when running the client's 'configure' script.

It fixes a compilation failure when squeezeplay_fab4 is built, and the
build system also has a local version of 'sdl-config' installed.

This change requires that client programs should run 'autoreconf' to
refresh their configure/makefile scripts. The Squeezeplay recipes in
the squeezeos poky build system do this automatically.

A) SDL

Takes up a fix to SDL's Automake macro to prevent it from using the
build system's locally installed version of 'sdl-config' when cross
compiling a client program.

In this circumstance, we should be using the version of 'sdl-config'
installed during the cross build process. The build will fail if the
build system's locally installed version is used.

Source:
sdl.m4: don't play with the PATH if cross-compiling
https://github.com/libsdl-org/SDL-1.2/commit/f7aab57364a9d3f608a7f6999b951dd6cc8c11f2

B) Squeezeplay

Remove 'acinclude.m4' from each of squeezeplay, squeezeplay_baby,
squeezeplay_jive, and squeezeplay_squeezeos.

These comprise a modified version of sdl.m4 that also dealt with the problem,
but are not now required as a result of the above fix.

Notably, this file was not included in squeezeplay_fab4, hence the
problem.