Squeezeplay-baby - Remove GNU Autotools generated build scripts
The poky/bitbake recipe in SqueezeOS is configured to run Gnu Autotools
each time a build is undertaken. All required build scripts are
regenerated at that time.

The files removed in this change are the products of earlier autotools
invocations. They are out of date, and will be regenerated on demand.