Squeezeplay-SqueezeOS - Remove GNU Autotools generated build scripts
The poky/bitbake recipe in SqueezeOS is configured to run Gnu Autotools
each time a build is undertaken. All required build scripts are
regenerated at that time.

Except as noted below, the files removed in this change are the products
of earlier autotools invocations. They are out of date, and will be
regenerated on demand.

The files Makefile.am~ and configure.in~ appear to be "left over" backup
files that do not participate in the build. They made their appearance as part
of refactoring change effected in 2008.

They are unnecessary and have also been removed.