Squeezeplay-fab4 - Remove GNU Autotools generated build scripts
The poky/bitbake recipe in SqueezeOS is configured to run Gnu Autotools
each time a build is undertaken. All required build scripts are
regenerated at that time.

The files removed in this change are the products of earlier autotools
invocations. They are out of date, and will be regenerated on demand.

Note that the file fab4_bsp.loT is a temporary file produced by the ltmain.sh
component of libtool, and should have been deleted or renamed when ltmain.sh
ran. We do not expect to see it remaining in the source tree.

It does not influence autotool operations - it will simply be overwritten and
removed.

(It was uploaded as part of the first commit back in 2008.)