SetupNetTestApplet - Fix "Unable to play the file type" popup
When running the Network speed test on some devices, a popup appears
"Problem: Unable to play the file type for teststream://test?rate=1000".

At this point, the Applet will immediately try to restart the speed test,
triggering a deluge of popups. It is almost impossible to back out of the
Applet (to stop the test) in these circumstances.

This is a consequence of the Applet's 'autostart' behaviour, which is
triggered whenever the speed test is not running.

This change adds an 'autostarted' flag to ensure that 'autostart' is
only activated once, when the user first starts the Applet. This resolves the
problem.

It also implements the following:

- Fix up the LMS query so that the list of allowed rates is requested only
when needed, rather than every second.

- Ensure that speed test rates values are of numeric type, particularly when
used as keys and values in Lua tables. Some failures were observed during
testing, when string types were unexpectedly encountered. This broke table
lookups.

- Add right button "help" to allow easy access to the help menu on a Touch.

Text display changes:

- Explain use of Centre/Go button to access help menu.

- Modify Help window, to indicate how different rates may be selected.